refactor: move CodeAuditResult schema into homeboy-audit-contract
Move the audit result value types (CodeAuditResult, AuditSummary,
ConventionReport, DirectoryConvention, DirectoryOutlier, Outlier,
Deviation, DuplicateGroup, CheckStatus) out of homeboy-core's code_audit
module into the shared homeboy-audit-contract crate.

These are the shared audit-output schema that both the audit engine
(producer) and the refactor engine (consumer, which also reconstructs
CodeAuditResult) depend on. Housing them in the contract removes another
class of cross-engine coupling and is a prerequisite for extracting the
audit engine into its own crate.

Pure serde value types only — timing (AuditTiming, with its core
phase_timing impls), AuditAnalysisContext, and audit execution behavior
stay in code_audit. All types are re-exported from their original
modules so existing code_audit::X paths resolve unchanged; no consumer
edits required. Bottoms out at AuditFinding/Finding, already in the
contract.

Refs #8425
