+++ crates/homeboy-audit-contract/src/result.rs
@@ -1,0 +1,150 @@
+//! Audit result value types — the shared schema that both the audit engine
+//! (which produces it) and the refactor engine (which consumes and reconstructs
+//! it) depend on. These are pure serde value types with no behavior; timing and
+//! internal analysis state stay in the audit engine.
+
+use serde::{Deserialize, Serialize};
+
+use crate::finding::Finding;
+use crate::finding_kind::AuditFinding;
+
+fn is_zero(value: &usize) -> bool {
+    *value == 0
+}
+
+/// Summary counts for the audit report.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct AuditSummary {
+    pub files_scanned: usize,
+    pub conventions_detected: usize,
+    #[serde(skip_serializing_if = "is_zero", default)]
+    pub outliers_found: usize,
+    /// Overall alignment score (0.0 = total chaos, 1.0 = perfect consistency).
+    /// Null when no files could be fingerprinted (score would be meaningless).
+    #[serde(skip_serializing_if = "Option::is_none", default)]
+    pub alignment_score: Option<f32>,
+    /// Source files found but not fingerprinted (no extension provides fingerprinting).
+    #[serde(skip_serializing_if = "is_zero", default)]
+    pub files_skipped: usize,
+    /// Warnings about the audit (e.g., unsupported file types).
+    #[serde(skip_serializing_if = "Vec::is_empty", default)]
+    pub warnings: Vec<String>,
+}
+
+/// Complete result of auditing a component's code conventions.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct CodeAuditResult {
+    pub component_id: String,
+    pub source_path: String,
+    pub summary: AuditSummary,
+    pub conventions: Vec<ConventionReport>,
+    #[serde(skip_serializing_if = "Vec::is_empty", default)]
+    pub directory_conventions: Vec<DirectoryConvention>,
+    pub findings: Vec<Finding>,
+    /// Grouped duplications for the fixer — each group has a canonical file and removal targets.
+    #[serde(skip_serializing_if = "Vec::is_empty", default)]
+    pub duplicate_groups: Vec<DuplicateGroup>,
+}
+
+/// A cross-directory convention: a pattern that sibling subdirectories share.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct DirectoryConvention {
+    /// Parent directory path (e.g., "inc/Abilities").
+    pub parent: String,
+    /// Expected methods that most subdirectories' conventions share.
+    pub expected_methods: Vec<String>,
+    /// Expected registrations that most subdirectories share.
+    #[serde(skip_serializing_if = "Vec::is_empty", default)]
+    pub expected_registrations: Vec<String>,
+    /// Subdirectories that conform.
+    pub conforming_dirs: Vec<String>,
+    /// Subdirectories that deviate.
+    #[serde(skip_serializing_if = "Vec::is_empty", default)]
+    pub outlier_dirs: Vec<DirectoryOutlier>,
+    /// How many subdirectories were analyzed.
+    pub total_dirs: usize,
+    /// Confidence score.
+    pub confidence: f32,
+}
+
+/// A subdirectory that deviates from the cross-directory convention.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct DirectoryOutlier {
+    /// Subdirectory name.
+    pub dir: String,
+    /// What's missing compared to sibling conventions.
+    pub missing_methods: Vec<String>,
+    /// Missing registrations.
+    #[serde(skip_serializing_if = "Vec::is_empty", default)]
+    pub missing_registrations: Vec<String>,
+}
+
+/// A convention as reported to the user (includes check status).
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ConventionReport {
+    pub name: String,
+    pub glob: String,
+    pub status: CheckStatus,
+    pub expected_methods: Vec<String>,
+    #[serde(skip_serializing_if = "Vec::is_empty", default)]
+    pub expected_registrations: Vec<String>,
+    #[serde(skip_serializing_if = "Vec::is_empty", default)]
+    pub expected_interfaces: Vec<String>,
+    #[serde(skip_serializing_if = "Option::is_none", default)]
+    pub expected_namespace: Option<String>,
+    #[serde(skip_serializing_if = "Vec::is_empty", default)]
+    pub expected_imports: Vec<String>,
+    pub conforming: Vec<String>,
+    #[serde(skip_serializing_if = "Vec::is_empty", default)]
+    pub outliers: Vec<Outlier>,
+    pub total_files: usize,
+    pub confidence: f32,
+}
+
+/// Status of a convention check.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum CheckStatus {
+    /// All files conform.
+    Clean,
+    /// Some files deviate.
+    Drift,
+    /// Less than half conform — convention may be wrong or split.
+    Fragmented,
+}
+
+/// A file that deviates from a convention.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Outlier {
+    /// Relative file path.
+    pub file: String,
+    /// Whether this outlier appears to be helper/utility drift rather than a real member.
+    #[serde(skip_serializing_if = "std::ops::Not::not", default)]
+    pub noisy: bool,
+    /// What's missing or different.
+    pub deviations: Vec<Deviation>,
+}
+
+/// A specific deviation from the convention.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Deviation {
+    /// What kind of deviation.
+    pub kind: AuditFinding,
+    /// Human-readable description.
+    pub description: String,
+    /// Suggested fix.
+    pub suggestion: String,
+}
+
+/// A group of files containing an identical function.
+///
+/// The fixer uses this to keep the canonical copy and remove the rest.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct DuplicateGroup {
+    /// The duplicated function name.
+    pub function_name: String,
+    /// File chosen to keep the function (canonical location).
+    pub canonical_file: String,
+    /// Files where the duplicate should be removed and replaced with an import.
+    pub remove_from: Vec<String>,
+}
